+++ resources/views/vote/event/show.blade.php
@@ -79,12 +79,16 @@
                                 <hr />
                                 <div>
                                     @if(Auth::check() && Auth::user()->isStaff())
-                                        {!! HTML::linkRoute('vote-event.edit', '編輯投票活動', $voteEvent->id, ['class' => 'btn btn-primary']) !!}
+                                        @if(!$voteEvent->isEnded())
+                                            {!! HTML::linkRoute('vote-event.edit', '編輯投票活動', $voteEvent->id, ['class' => 'btn btn-primary']) !!}
+                                        @endif
                                         {!! HTML::linkRoute('vote-event.index', '返回投票活動列表', [], ['class' => 'btn btn-default']) !!}
-                                        {!! Form::open(['route' => ['vote-event.destroy', $voteEvent->id], 'style' => 'display: inline', 'method' => 'DELETE',
-                                        'onSubmit' => "return confirm('確定要刪除投票活動嗎？');"]) !!}
-                                        {!! Form::submit('刪除', ['class' => 'btn btn-danger']) !!}
-                                        {!! Form::close() !!}
+                                        @if(!$voteEvent->isStarted())
+                                            {!! Form::open(['route' => ['vote-event.destroy', $voteEvent->id], 'style' => 'display: inline', 'method' => 'DELETE',
+                                            'onSubmit' => "return confirm('確定要刪除投票活動嗎？');"]) !!}
+                                            {!! Form::submit('刪除', ['class' => 'btn btn-danger']) !!}
+                                            {!! Form::close() !!}
+                                        @endif
                                         @if($voteEvent->isInProgress())
                                             {!! HTML::linkRoute('vote.vote', '進入投票頁面', ['id' => $voteEvent->id], ['class' => 'btn btn-primary']) !!}
                                         @endif
